Time: 2 ms (41.69%) | Memory: 7.8 MB (42.04%) - LeetSync

+++ 171-excel-sheet-column-number/excel-sheet-column-number.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int titleToNumber(string columnTitle) {
-        long long int res = 0;
+        int res = 0;
         for(int i = 0;i < columnTitle.size(); i++)
         {
-            long long int temp = ((long int) columnTitle[i] - 'A') + 1;
+            int temp = ((long int) columnTitle[i] - 'A') + 1;
             res = res * 26 + temp;
         }
         return res;
